pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "groundguard"
version = "0.1.0"
description = "LLM-agnostic deterministic assert layer for AI agent outputs"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "litellm>=1.0",
    "pydantic>=2.0",
    "rank-bm25>=0.2",
    "numpy>=1.24",
    "rapidfuzz>=3.0",
]

[project.optional-dependencies]
loaders = [
    "pypdf>=3.0",
    "python-docx>=1.0",
]
langchain = [
    "langchain-core>=0.1",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-mock>=3.10",
    "pytest-cov>=4.0",
    "pytest-timeout>=2.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "llm: marks tests as real LLM integration tests (deselect with '-m not llm')",
    "loaders: marks tests requiring optional loaders extras",
    "langchain: marks tests requiring optional langchain extras",
    "compat: marks tests as multi-model compatibility smoke tests (parametrized across all compat models)",
]
addopts = "--cov=groundguard --cov-report=term-missing"
timeout = 300

[tool.coverage.run]
source = ["groundguard"]
omit = ["*/tests/*"]

[tool.hatch.build.targets.wheel]
packages = ["groundguard"]
